requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlmodel==0.0.18
aiosqlite==0.20.0

# HTTP client
httpx==0.27.0

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.1

# Auth
passlib[bcrypt]==1.7.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
respx==0.21.1
